inputView -> resultView 로 enterResultUser 메소드 이동
