+++ src/test/java/nextstep/ladder/view/ResultViewTest.java
@@ -1,7 +1,9 @@
 package nextstep.ladder.view;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -16,6 +18,7 @@ import nextstep.ladder.domain.ImprovingPoint;
 import nextstep.ladder.domain.Ladder;
 import nextstep.ladder.domain.LadderLine;
 import nextstep.ladder.domain.ResultsTest;
+import nextstep.ladder.domain.User;
 import nextstep.ladder.domain.Users;
 import nextstep.ladder.domain.UsersTest;
 
@@ -52,6 +55,26 @@ class ResultViewTest {
         ladder = new Ladder(lineList);
         gameInfo = new GameInfo(UsersTest.USERS, ResultsTest.RESULTS);
         resultUsers.generateResultsForAllPlayers(gameInfo, ladder);
+    }
+
+    @DisplayName("결과를 보고싶은 사람을 입력받을 경우 유저에 포함되는지 확인해본다. "
+                 + "아닐경우 exception 발생시킨다.")
+    @Test
+    void validateResultUser() {
+        Users users = new Users(new ArrayList<>(
+                Stream.of(new User("pobi"),
+                          new User("honux"),
+                          new User("crong"),
+                          new User("jk")
+                )
+                      .collect(Collectors.toList())));
+
+        String userName = "soo";
+
+        assertThatIllegalArgumentException().isThrownBy(() -> {
+            resultView.validateResultUser(users, userName);
+        });
+
     }
 
     @DisplayName("유저 이름을 출력한다.")
